// backend/src/domain/enums.ts
export enum ApartmentStatus {
  New = 'New',
  Launch = 'Launch',
  ReadyToMove = 'Ready to Move',
  NearDelivery = 'Near Delivery',
}

export enum SaleType {
  DeveloperSale = 'Developer Sale',
  Resale = 'Resale',
  NawyNow = 'Nawy Now',
}

export enum Finishing {
  FullyFinished = 'Fully Finished',
  Unfinished = 'Unfinished',
  SemiFinished = 'Semi-Finished',
}

export enum InstallmentFrequency {
  Monthly = 'Monthly',
  Quarterly = 'Quarterly',
  SemiAnnually = 'Semi-Annually',
}

export enum Currency {
  EGP = 'EGP',
}